+++ src/reprepbuild/builtin/__init__.py
@@ -21,7 +21,7 @@ from .check_hrefs import check_hrefs
 from .latex import latex, latex_diff, latex_flat
 from .mupdf import pdf_add_notes, pdf_merge, pdf_nup
 from .python_script import python_script
-from .transform import convert_odf_pdf, convert_svg_pdf, copy, pdf_raster, render
+from .transform import convert_odf_pdf, convert_svg_pdf, copy, markdown_pdf, pdf_raster, render
 from .zip import repro_zip, repro_zip_latex
 
 __all__ = ("get_commands",)
@@ -40,6 +40,7 @@ def get_commands():
         convert_odf_pdf,
         convert_svg_pdf,
         copy,
+        markdown_pdf,
         pdf_raster,
         render,
         repro_zip,